+++ CHANGELOG.md
@@ -85,3 +85,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [t-81] Ajout des pages dans middleware [#81](https://trello.com/c/BnwXnjKD/81-t-81-ajout-des-pages-dans-middleware)
 - [t-82] Ajout du composant pour créer un post [#82](https://trello.com/c/3YzEFTKc/82-t-82-ajout-du-composant-pour-cr%C3%A9er-un-post)
 - [t-83] Mise en place du feed [#83](https://trello.com/c/JU9GLCah/83-t-83-mise-en-place-du-feed)
+- [t-84] Afficher les commentaires d'un post dans PostCard.tsx [#84](https://trello.com/c/Lkhs6ulR/84-t-84-afficher-les-commentaires-dun-post-dans-postcardtsx)